requirements.txt
streamlit
pandas
openai
sqlalchemy
sqlite3 
plotly
requests
langchain
langchain-openai
langchain-community
langgraph
google-cloud-bigquery
